+++ config/inkforge-editor.php
@@ -37,7 +37,7 @@ return [
         ['undo', 'redo'],
         ['blockFormat', 'fontFamily', 'fontSize'],
         ['bold', 'italic', 'underline', 'strike', 'superscript', 'subscript'],
-        ['forecolor', 'backcolor'],
+        ['forecolor', 'backcolor', 'removeFormat'],
         ['alignLeft', 'alignCenter', 'alignRight', 'alignJustify'],
         ['bulletList', 'orderedList', 'checklist', 'indent', 'outdent'],
         ['link', 'unlink', 'image', 'video', 'audio', 'table', 'hr'],